Converter parser a folder of files

+++ RQ2_Kotlin/src/main/kotlin/ch/uzh/ifi/seal/smr/reconfigure/Converter.kt
@@ -4,15 +4,21 @@ import ch.uzh.ifi.seal.bencher.jmhResults.JMHResultTransformer
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
+import java.nio.file.Paths
 
 fun main() {
-    val inputFile = File("D:\\projects\\apache#logging-log4j2\\log4j-perf\\target\\out-log4j2-2-local.json")
-    val outputFile = File("D:\\projects\\apache#logging-log4j2\\log4j-perf\\target\\out-log4j2-2-local.csv")
+    val inputFolder = File("D:\\rq2\\results\\results-jenetics")
+    val outputFolder = File("D:\\rq2\\results\\csv-jenetics")
+    inputFolder.walk().forEach {inputFile ->
+        if(inputFile.isFile){
+            val outputFile = Paths.get(outputFolder.absolutePath, inputFile.nameWithoutExtension + ".csv").toFile()
 
-    val parser = JMHResultTransformer(inStream = FileInputStream(inputFile), outStream = FileOutputStream(outputFile), trial = 0, commit = "", project = "", instance = "")
-    val error = parser.execute()
-    if (error.isDefined()) {
-        throw RuntimeException("Execution failed with '${error.get()}'")
+            val parser = JMHResultTransformer(inStream = FileInputStream(inputFile), outStream = FileOutputStream(outputFile), trial = 0, commit = "", project = "", instance = "")
+            val error = parser.execute()
+            if (error.isDefined()) {
+                throw RuntimeException("Execution failed with '${error.get()}'")
+            }
+        }
     }
 }
